Bug 560284 - [win32][Dark] Table header delimiters color is bad for dark theme
This patch introduces a way to configure the color.

The configuration is done via `Table.setData()` to avoid the need for
reflection in user's cross-platform code.

I have considered using border color for lines, because Windows already
seems to do a good job, both for dark and light themes. However, I
decided against it after seeing that in default light theme Windows uses
different colors for border and header lines.

Change-Id: I9a1dec43d6081097bb0db0a47a0f381163ba5b32
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>