Bug 567422 - [Win32][DND] Drag and Drop can crash if target keep reference to COM object
The aggressive release of all COM objects after the drag/drop
operation is done can be problematic if the drop target keeps a
reference to one of the COM objects for later use. When the target later
tries to release (or access) the already disposed COM object a crash
from invalid memory access is likely.

Instead if should trust that the target will eventually release the
references it still holds.

For the case that the target will request the data again after DND is
finished the last requested data is cached. Note: sending a DragSetData
event before DragStart or after DragEnd is considered a bug.

Change-Id: I48e88a58e46111a3879e071d56abb79ac75114a9
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>