Bug 570331 - [GTK] Combo can't be interacted with after .setRedraw(false) + .setRedraw(true)
This patch fixes multiple problems at once:
1) Combo.setRedraw(false) + Combo.setRedraw(true) made it non-responsive
2) Combo.redraw() didn't work
3) GC based on Combo covered wrong part of Control

The problem was that SWT searched for a Combo's subwindow, which GTK
creates with GDK_INPUT_ONLY (that is, not suitable for any painting),
and actually tried to use it for painting.

It was probably an oversight from the very beginning, resulting from a
series of workarounds where SWT tried to separate windows used for
different APIs.

See also Bug 561444 and its commit message.

Change-Id: I24400d175f31898723bb1cdc6313934d3903bf2c
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>