Bug 571740 - [GTK] StyledText: IllegalArgumentException on dead keys
The GTK's bug is in `gtk_im_context_simple_get_preedit_string()`, which
incorrectly returns `len`, which is expressed in bytes instead of
characters.

Note that on GNOME, SWT forces a different IM called "ibus" via
Bug 546349 and the bug is not seen there. The bug also won't trigger on
systems that are configured to use any other IM.

Change-Id: I4d24a62e9aa34e0ee6fe2d3ec1395f97aa89dfd6
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>