Issue #351: Remove early return in `Widget.wmKeyUp()`
The code existed since the first commit in the repository, and it's not
clear why it was added. I understand that it helped with dead keys, but
it was reworked multiple times since then.

This code stands in the way of various refactorings.
Also, when present, it suppressed `SWT.KeyUp` in the next commit.

I'm rather confident that this is a valid change:
* unit tests in `Test_org_eclipse_swt_events_KeyEvent` pass
  they test US layout very well, and also test some dead keys.
* I manually tested IME (Japanese) in the next commit

Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>