#450/Bug 578929 - Control.setFocus brings windows to front
If the application is in the background and the GUI is updated,
especially a different control is focused, it might happen that the
shell will be brought to front.

This patch should fix that if you invoke

 display.setData(Display.ACTIVATE_SHELL_ON_FORCE_FOCUS, false);

after having created the display instance. If you want an application
window be brought to front, use shell.forceActive().