#450/Bug 578929 - Control.setFocus brings windows to front
This change introduces the new method `control.setFocus(boolean
activateShell)` as a replacement for the now deprecated
`control.setFocus()`. The deprecated method will delegate to the new
method with the default for `activateShell` configured by the system
property `org.eclipse.swt.internal.activateShellOnForceFocus`.

The deprecated method `control.setFocus()` has been made `final` to
force a compiler error if it is overridden. In that case the developer
has to override `control.setFocus(boolean)` instead, passing the
`activateShell` parameter (see other changed controls in this commit).