Issue 0930: [GTK] Eclipse crashes if gnome-orca is enabled
It crashed because of a missing call to gtk_accessible_set_widget.

According to the documentation https://docs.gtk.org/gtk3/method.Accessible.set_widget.html it has to be called when the widget is unregistered as well.

Quote from documentation: "accessible will not hold a reference to widget. It is the caller’s responsibility to ensure that when widget is destroyed, the widget is unset by calling this function again with widget set to NULL."

Signed-off-by: Rene Windegger <rene.windegger@syntevo.com>